requirements.txt
numpy==1.17.4
setuptools==57.5.0
psycopg2>=2.8
Django>=3.0,<4.0
djangorestframework==3.12.4
gunicorn==20.1.0
dj-database-url==0.5.0